Change expected hostname env var name
The HOSTNAME env is a common unix variable, and is set for things like
containers in a Kubernetes cluster. This means that if you are using a
file for config, the hostname in the config will be ignored and replaced
with the machine or container name. This was not the intention.

We should clarify the name of the environment variable we are looking
for to `TARGET_HOSTNAME` to avoid this conflict.